Add files via upload
we have the data for the 100 top-rated movies from the past decade along with various pieces of information about the movie, its actors, and the voters who have rated these movies online. In this assignment, we tried to find some interesting insights into these movies and their voters, using Python.